server/requirements.txt
fastapi
uvicorn[standard]
numpy
Pillow
python-multipart
pydantic-settings
torch
torchvision
ultralytics